requirements.txt
tweepy
requests
pillow
pip
aggdraw